requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==2.0.1
click==8.0.4
colorama==0.4.4
fastapi==0.75.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
pycodestyle==2.10.0
pydantic==1.9.0
python-dotenv==0.19.2
slack-bolt==1.11.2
slack-sdk==3.13.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5
zipp==3.7.0
